Added log in/create user stages
Add: #1
