+++ database.txt
@@ -1,0 +1,2 @@
+admin:123456,null
+cng:98765,cars
