+++ nbproject/private/private.xml
@@ -2,6 +2,9 @@
 <project-private xmlns="http://www.netbeans.org/ns/project-private/1">
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
-        <group/>
+        <group>
+            <file>file:/D:/resume/Java%20Work%20Samples/web%20projects/conference_gui/src/conference_gui/ClearButtonEvents.java</file>
+            <file>file:/D:/resume/Java%20Work%20Samples/web%20projects/conference_gui/src/conference_gui/Conference_gui.java</file>
+        </group>
     </open-files>
 </project-private>
